# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Metro API
  description: The client-server API provides an opportunity to view the interval of movement of trains of the Minsk metro in accordance with the current time
  contact:
    name: Alexey Burak
    url: https://www.linkedin.com/in/alexey-burak/
    email: burakalexey@yahoo.com
  license:
    name: Apache 2.0
    url: https://github.com/alexeyburak/metro/blob/main/LICENSE
  version: 1.0.0
externalDocs:
  description: API specification
  url: https://github.com/alexeyburak/metro/blob/main/docs/Specification-metro.md
servers:
  - url: http://localhost:3000/api
    description: Development server
tags:
  - name: Lines
    description: API for managing lines
paths:
  /lines/red:
    get:
      tags:
        - Lines
      summary: Get red line interval
      operationId: redLineInfo
      responses:
        '200':
          description: Successfully get interval
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LineIntervalResponse'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
        '404':
          description: Line is not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
  /lines/green:
    get:
      tags:
        - Lines
      summary: Get green line interval
      operationId: greenLineInfo
      responses:
        '200':
          description: Successfully get interval
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LineIntervalResponse'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
        '404':
          description: Line is not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
  /lines/blue:
    get:
      tags:
        - Lines
      summary: Get blue line interval
      operationId: blueLineInfo
      responses:
        '200':
          description: Successfully get interval
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LineIntervalResponse'
        '400':
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
        '404':
          description: Line is not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiException'
components:
  schemas:
    ApiException:
      type: object
      properties:
        message:
          type: string
        httpStatus:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
        phrase:
          type: string
    LineIntervalResponse:
      type: object
      properties:
        name:
          type: string
        time:
          type: number
          format: double
        allTime:
          type: integer
          format: int32