# src/main/java/com/burak/metro/dto/LineIntervalResponse.java
package com.burak.metro.dto;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;
import lombok.experimental.FieldDefaults;

import static lombok.AccessLevel.PRIVATE;

@Getter
@Setter
@Builder
@FieldDefaults(level = PRIVATE)
public class LineIntervalResponse {
    String name;
    Double time;
    Integer allTime;
}